+++ Private/MSCatalog/Invoke-CatalogRequest.ps1
@@ -64,8 +64,14 @@ function Invoke-CatalogRequest {
                 throw "The catalog.update.microsoft.com site has encountered an error. Please try again later."
             }
             else {
-                $HtmlDoc | Out-File -FilePath "$env:TEMP\$((Get-Date).ToString('yyyy-MM-dd-HHmmss')) OSDCloud Microsoft Catalog.html" -Encoding "UTF8"
-                [MsUpCatResponse]::new($HtmlDoc)
+                #Changed for Issue 127
+                if($null -eq $HtmlDoc.GetElementbyId("ctl00_catalogBody_updateMatches")){ 
+                    return $null 
+                }
+                else {
+                    $HtmlDoc | Out-File -FilePath "$env:TEMP\$((Get-Date).ToString('yyyy-MM-dd-HHmmss')) OSDCloud Microsoft Catalog.html" -Encoding "UTF8"
+                    [MsUpCatResponse]::new($HtmlDoc)
+                }
             }
         } 
         else {